% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_apply_formats.R
\name{apply_formats}
\alias{apply_formats}
\title{apply_formats}
\usage{
apply_formats(., specs_data, view_name)
}
\arguments{
\item{.}{current data frame}

\item{specs_data}{a data frame with expected variables including (column_name as the targerted variable name, format_name as the desired format
label name and view_name as the current view being cleaned and formatted)}

\item{view_name}{the name of the view being cleaned (i.e. bv_exits)}
}
\value{
a data frame with DWH lookup variables formatted from the meta data
}
\description{
Function that takes in the current data frame, a specs data frame (with at least the column_name as the DWH current variable name,
the format label name and the view_name) and applies all of the label formatting to each variable
}
