% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05b_assign_value_labels.R
\name{assign_value_labels}
\alias{assign_value_labels}
\title{assign_value_labels}
\usage{
assign_value_labels(., value_labels_df)
}
\arguments{
\item{.}{current data frame}

\item{value_labels_df}{value labels data frame (expected columns are the new_varname containing the name of the variable and new_valuelabel
containing the desired desired format label name of the format object expected to be stored in the working environment)}
}
\value{
a data frame with targeted categorical variables labelled with their assigned value format
}
\description{
Function to assign value labels to the new variable names (designated as new_varname) and the desired format label name
(designated by new_valuelabel)
}
