% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_assign_var_labels.R
\name{assign_var_labels}
\alias{assign_var_labels}
\title{assign_var_labels}
\usage{
assign_var_labels(., var_labels_df)
}
\arguments{
\item{.}{current data frame}

\item{var_labels_df}{variable labels data frame (expected columns are the new_varname containing the name of the variable and new_varlabel
containing the desired label of the variable)}
}
\value{
data frame with labelled variables
}
\description{
Function to assign variable labels to the new variable names (designated as new_varname) and the desired variable
label (designated by new_varlabel)
}
