% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_clean_date_vars.R
\name{clean_date_vars}
\alias{clean_date_vars}
\title{clean_date_vars}
\usage{
clean_date_vars(dt_var)
}
\arguments{
\item{dt_var}{supplied date variable(s)}
}
\value{
a list with a formatted date variable (dt) and the imputation status (imp_status) of the date variable. Also returns a format
object (date_impf) to the current working environment
}
\description{
Function that cleans the date variable formatting and creates two final variables, including a formatted date variable (dt)
and the imputation status (imp_status) of the date variable
}
