% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_create_common_formats.R
\name{create_common_formats}
\alias{create_common_formats}
\title{create_common_formats}
\usage{
create_common_formats()
}
\value{
various formats to be used for labeling (i.e. race7_f, race4_f, smoker3_f, female_male_f, etc.)
}
\description{
Function that creates commonly used formats for race (4 categories, 7 categories) 3 smoker categories, female/male common variable
as well as some other common formats
}
