% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_initiate_submodule.R
\name{initiate_submodule}
\alias{initiate_submodule}
\title{initiate_submodule}
\usage{
initiate_submodule(submodule_to_add, submodule_git_path)
}
\arguments{
\item{submodule_to_add}{name of submodule to add to parent repository}

\item{submodule_git_path}{github path to submodule (i.e. https://github.com/Corrona-IT/EDW_UAT_Integration_Setup)}
}
\value{
an added submodule in the working directory repository
}
\description{
initiate_submodule
}
