% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_remove_submodule.R
\name{remove_submodule}
\alias{remove_submodule}
\title{remove_submodule}
\usage{
remove_submodule(submodule_to_rm)
}
\arguments{
\item{submodule_to_rm}{name of the sumodule you want to remove}
}
\value{
does not return anything
}
\description{
Function to remove a git submodule from a repository
}
