% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rename_edw_vars.R
\name{rename_edw_vars}
\alias{rename_edw_vars}
\title{rename_edw_vars}
\usage{
rename_edw_vars(., rename_vars_df)
}
\arguments{
\item{.}{current data frame}

\item{rename_vars_df}{variable name mapping data frame (expected columns are the old_varname containing the name of the EDW DWH variable name to be renamed and
the new_varname containing the name of the final desired variable name)}
}
\value{
a data frame with renamed EDW DWH variables renamed to their desired final analytical name
}
\description{
Function that renames all designated EDW DWH variables (old_varname) to the desired final analyctical variable name (new_varname)
}
